Add CacheFSRuntime CRD definitions and controller implementations
Cachefs is a distributed filesystem that cache & accelerate any given
POSIX directory with distributed cache.

And ThinRuntime was implemented to make general storage like oss/nas
easier to integrate with fluid, but thinruntime doesn't have the ability
to cache & accelerate data access on thinruntime.

So CacheFSRuntime is add as a general way to accelerate any storage from
thinruntime or any given dir.

Fixes: #3674
Signed-off-by: zyfjeff <zyfjeff@linux.alibaba.com>
Signed-off-by: Eryu Guan <eguan@linux.alibaba.com>